+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This release accompanies the release of our paper "Global continuation as a comp
 
 ## New features
 
+- `basin_entropy` function has been extended to work for sampled basins by using
+  nearest neighbor searches and KD trees.
+- `basin_entropy` is now also optionally estimated by `StabilityQuantifierAccumulator`.
 - New function `hilbert_pcurve`. It conveniently uses Hilbert curves to create a parameter curve efficiently spanning a multidimensional space. To be used with `global_continuation`.
 - `StabilityQuantifiersAccumulator` allows for an `extras` input to calculate arbitrary
   additional quantities after sampling the state space.